% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_objects.R
\name{AwsAccessKey}
\alias{AwsAccessKey}
\title{AwsAccessKey Object}
\usage{
AwsAccessKey(accessKeyId = NULL, secretAccessKey = NULL)
}
\arguments{
\item{accessKeyId}{AWS access key ID}

\item{secretAccessKey}{AWS secret access key}
}
\value{
AwsAccessKey object
}
\description{
AwsAccessKey Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
AWS access key (see [AWS Security Credentials](http://docs.aws.amazon.com/general/latest/gr/aws-security-credentials.html)).
}
