% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_objects.R
\name{AwsS3Data}
\alias{AwsS3Data}
\title{AwsS3Data Object}
\usage{
AwsS3Data(bucketName = NULL, awsAccessKey = NULL)
}
\arguments{
\item{bucketName}{S3 Bucket name (see [Creating a bucket](http://docs}

\item{awsAccessKey}{AWS access key used to sign the API requests to the AWS S3 bucket}
}
\value{
AwsS3Data object
}
\description{
AwsS3Data Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An AwsS3Data can be a data source, but not a data sink. In an AwsS3Data, an object's name is the S3 object's key name.
}
