% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_objects.R
\name{ErrorLogEntry}
\alias{ErrorLogEntry}
\title{ErrorLogEntry Object}
\usage{
ErrorLogEntry(url = NULL, errorDetails = NULL)
}
\arguments{
\item{url}{A URL that refers to the target (a data source, a data sink, or an object) with which the error is associated}

\item{errorDetails}{A list of messages that carry the error details}
}
\value{
ErrorLogEntry object
}
\description{
ErrorLogEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An entry describing an error that has occurred.
}
