% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_objects.R
\name{ErrorSummary}
\alias{ErrorSummary}
\title{ErrorSummary Object}
\usage{
ErrorSummary(errorCode = NULL, errorCount = NULL, errorLogEntries = NULL)
}
\arguments{
\item{errorCode}{Required}

\item{errorCount}{Count of this type of error}

\item{errorLogEntries}{Error samples}
}
\value{
ErrorSummary object
}
\description{
ErrorSummary Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A summary of errors by error code, plus a count and sample error log entries.
}
