% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_objects.R
\name{GcsData}
\alias{GcsData}
\title{GcsData Object}
\usage{
GcsData(bucketName = NULL)
}
\arguments{
\item{bucketName}{Google Cloud Storage bucket name (see [Bucket Name Requirements](https://cloud}
}
\value{
GcsData object
}
\description{
GcsData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
In a GcsData, an object's name is the Google Cloud Storage object's name and its `lastModificationTime` refers to the object's updated time, which changes when the content or the metadata of the object is updated.
}
