% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_objects.R
\name{GoogleServiceAccount}
\alias{GoogleServiceAccount}
\title{Google Storage Transfer API Objects 
Transfers data from external data sources to a Google Cloud Storage bucket or between Google Cloud Storage buckets.}
\usage{
GoogleServiceAccount(accountEmail = NULL)
}
\arguments{
\item{accountEmail}{Required}
}
\value{
GoogleServiceAccount object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:58:43
filename: /Users/mark/dev/R/autoGoogleAPI/googlestoragetransferv1.auto/R/storagetransfer_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
GoogleServiceAccount Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Google service account
}
