% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_objects.R
\name{ListTransferJobsResponse}
\alias{ListTransferJobsResponse}
\title{ListTransferJobsResponse Object}
\usage{
ListTransferJobsResponse(transferJobs = NULL, nextPageToken = NULL)
}
\arguments{
\item{transferJobs}{A list of transfer jobs}

\item{nextPageToken}{The list next page token}
}
\value{
ListTransferJobsResponse object
}
\description{
ListTransferJobsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response from ListTransferJobs.
}
