% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_objects.R
\name{ObjectConditions}
\alias{ObjectConditions}
\title{ObjectConditions Object}
\usage{
ObjectConditions(minTimeElapsedSinceLastModification = NULL,
  maxTimeElapsedSinceLastModification = NULL, includePrefixes = NULL,
  excludePrefixes = NULL)
}
\arguments{
\item{minTimeElapsedSinceLastModification}{If unspecified, `minTimeElapsedSinceLastModification` takes a zero value and `maxTimeElapsedSinceLastModification` takes the maximum possible value of Duration}

\item{maxTimeElapsedSinceLastModification}{`maxTimeElapsedSinceLastModification` is the complement to `minTimeElapsedSinceLastModification`}

\item{includePrefixes}{If `includePrefixes` is specified, objects that satisfy the object conditions must have names that start with one of the `includePrefixes` and that do not start with any of the `excludePrefixes`}

\item{excludePrefixes}{`excludePrefixes` must follow the requirements described for `includePrefixes`}
}
\value{
ObjectConditions object
}
\description{
ObjectConditions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Conditions that determine which objects will be transferred.
}

