% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_objects.R
\name{PauseTransferOperationRequest}
\alias{PauseTransferOperationRequest}
\title{PauseTransferOperationRequest Object}
\usage{
PauseTransferOperationRequest()
}
\value{
PauseTransferOperationRequest object
}
\description{
PauseTransferOperationRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request passed to PauseTransferOperation.
}
\seealso{
Other PauseTransferOperationRequest functions: \code{\link{transferOperations.pause}}
}
