% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_objects.R
\name{ResumeTransferOperationRequest}
\alias{ResumeTransferOperationRequest}
\title{ResumeTransferOperationRequest Object}
\usage{
ResumeTransferOperationRequest()
}
\value{
ResumeTransferOperationRequest object
}
\description{
ResumeTransferOperationRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request passed to ResumeTransferOperation.
}
\seealso{
Other ResumeTransferOperationRequest functions: \code{\link{transferOperations.resume}}
}

