% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_objects.R
\name{Schedule}
\alias{Schedule}
\title{Schedule Object}
\usage{
Schedule(scheduleStartDate = NULL, scheduleEndDate = NULL,
  startTimeOfDay = NULL)
}
\arguments{
\item{scheduleStartDate}{The first day the recurring transfer is scheduled to run}

\item{scheduleEndDate}{The last day the recurring transfer will be run}

\item{startTimeOfDay}{The time in UTC at which the transfer will be scheduled to start in a day}
}
\value{
Schedule object
}
\description{
Schedule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Transfers can be scheduled to recur or to run just once.
}

