% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_functions.R
\docType{package}
\name{storagetransfer_googleAuthR}
\alias{storagetransfer_googleAuthR}
\alias{storagetransfer_googleAuthR-package}
\title{Google Storage Transfer API
Transfers data from external data sources to a Google Cloud Storage bucket or between Google Cloud Storage buckets.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:58:43
filename: /Users/mark/dev/R/autoGoogleAPI/googlestoragetransferv1.auto/R/storagetransfer_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}
}
