% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_objects.R
\name{TransferCounters}
\alias{TransferCounters}
\title{TransferCounters Object}
\usage{
TransferCounters(objectsFoundFromSource = NULL, bytesFoundFromSource = NULL,
  objectsFoundOnlyFromSink = NULL, bytesFoundOnlyFromSink = NULL,
  objectsFromSourceSkippedBySync = NULL,
  bytesFromSourceSkippedBySync = NULL, objectsCopiedToSink = NULL,
  bytesCopiedToSink = NULL, objectsDeletedFromSource = NULL,
  bytesDeletedFromSource = NULL, objectsDeletedFromSink = NULL,
  bytesDeletedFromSink = NULL, objectsFromSourceFailed = NULL,
  bytesFromSourceFailed = NULL, objectsFailedToDeleteFromSink = NULL,
  bytesFailedToDeleteFromSink = NULL)
}
\arguments{
\item{objectsFoundFromSource}{Objects found in the data source that are scheduled to be transferred, which will be copied, excluded based on conditions, or skipped due to failures}

\item{bytesFoundFromSource}{Bytes found in the data source that are scheduled to be transferred, which will be copied, excluded based on conditions, or skipped due to failures}

\item{objectsFoundOnlyFromSink}{Objects found only in the data sink that are scheduled to be deleted}

\item{bytesFoundOnlyFromSink}{Bytes found only in the data sink that are scheduled to be deleted}

\item{objectsFromSourceSkippedBySync}{Objects in the data source that are not transferred because they already exist in the data sink}

\item{bytesFromSourceSkippedBySync}{Bytes in the data source that are not transferred because they already exist in the data sink}

\item{objectsCopiedToSink}{Objects that are copied to the data sink}

\item{bytesCopiedToSink}{Bytes that are copied to the data sink}

\item{objectsDeletedFromSource}{Objects that are deleted from the data source}

\item{bytesDeletedFromSource}{Bytes that are deleted from the data source}

\item{objectsDeletedFromSink}{Objects that are deleted from the data sink}

\item{bytesDeletedFromSink}{Bytes that are deleted from the data sink}

\item{objectsFromSourceFailed}{Objects in the data source that failed during the transfer}

\item{bytesFromSourceFailed}{Bytes in the data source that failed during the transfer}

\item{objectsFailedToDeleteFromSink}{Objects that failed to be deleted from the data sink}

\item{bytesFailedToDeleteFromSink}{Bytes that failed to be deleted from the data sink}
}
\value{
TransferCounters object
}
\description{
TransferCounters Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A collection of counters that report the progress of a transfer operation.
}
