% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_objects.R
\name{TransferJob}
\alias{TransferJob}
\title{TransferJob Object}
\usage{
TransferJob(name = NULL, description = NULL, projectId = NULL,
  transferSpec = NULL, schedule = NULL, status = NULL,
  creationTime = NULL, lastModificationTime = NULL, deletionTime = NULL)
}
\arguments{
\item{name}{A globally unique name assigned by Storage Transfer Service when the job is created}

\item{description}{A description provided by the user for the job}

\item{projectId}{The ID of the Google Developers Console project that owns the job}

\item{transferSpec}{Transfer specification}

\item{schedule}{Schedule specification}

\item{status}{Status of the job}

\item{creationTime}{This field cannot be changed by user requests}

\item{lastModificationTime}{This field cannot be changed by user requests}

\item{deletionTime}{This field cannot be changed by user requests}
}
\value{
TransferJob object
}
\description{
TransferJob Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This resource represents the configuration of a transfer job that runs periodically.
}
\seealso{
Other TransferJob functions: \code{\link{transferJobs.create}}
}
