% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_functions.R
\name{transferJobs.create}
\alias{transferJobs.create}
\title{Creates a transfer job that runs periodically.}
\usage{
transferJobs.create(TransferJob)
}
\arguments{
\item{TransferJob}{The \link{TransferJob} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/storage/transfer}{Google Documentation}

Other TransferJob functions: \code{\link{TransferJob}}
}
