% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_functions.R
\name{transferJobs.list}
\alias{transferJobs.list}
\title{Lists transfer jobs.}
\usage{
transferJobs.list(filter = NULL, pageSize = NULL, pageToken = NULL)
}
\arguments{
\item{pageSize}{The list page size}

\item{pageToken}{The list page token}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/storage/transfer}{Google Documentation}
}

