% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_functions.R
\name{transferJobs.patch}
\alias{transferJobs.patch}
\title{Updates a transfer job. Updating a job's transfer spec does not affect transfer operations that are running already. Updating the scheduling of a job is not allowed.}
\usage{
transferJobs.patch(UpdateTransferJobRequest, jobName)
}
\arguments{
\item{UpdateTransferJobRequest}{The \link{UpdateTransferJobRequest} object to pass to this method}

\item{jobName}{The name of job to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/storage/transfer}{Google Documentation}

Other UpdateTransferJobRequest functions: \code{\link{UpdateTransferJobRequest}}
}

