% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_objects.R
\name{TransferOperation}
\alias{TransferOperation}
\title{TransferOperation Object}
\usage{
TransferOperation(name = NULL, projectId = NULL, transferSpec = NULL,
  startTime = NULL, endTime = NULL, status = NULL, counters = NULL,
  errorBreakdowns = NULL, transferJobName = NULL)
}
\arguments{
\item{name}{A globally unique ID assigned by the system}

\item{projectId}{The ID of the Google Developers Console project that owns the operation}

\item{transferSpec}{Transfer specification}

\item{startTime}{Start time of this transfer execution}

\item{endTime}{End time of this transfer execution}

\item{status}{Status of the transfer operation}

\item{counters}{Information about the progress of the transfer operation}

\item{errorBreakdowns}{Summarizes errors encountered with sample error log entries}

\item{transferJobName}{The name of the transfer job that triggers this transfer operation}
}
\value{
TransferOperation object
}
\description{
TransferOperation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A description of the execution of a transfer.
}
