% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_functions.R
\name{transferOperations.cancel}
\alias{transferOperations.cancel}
\title{Cancels a transfer. Use the get method to check whether the cancellation succeeded or whether the operation completed despite cancellation.}
\usage{
transferOperations.cancel(name)
}
\arguments{
\item{name}{The name of the operation resource to be cancelled}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/storage/transfer}{Google Documentation}
}
