% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_functions.R
\name{transferOperations.pause}
\alias{transferOperations.pause}
\title{Pauses a transfer operation.}
\usage{
transferOperations.pause(PauseTransferOperationRequest, name)
}
\arguments{
\item{PauseTransferOperationRequest}{The \link{PauseTransferOperationRequest} object to pass to this method}

\item{name}{The name of the transfer operation}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/storage/transfer}{Google Documentation}

Other PauseTransferOperationRequest functions: \code{\link{PauseTransferOperationRequest}}
}
