% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_functions.R
\name{transferOperations.resume}
\alias{transferOperations.resume}
\title{Resumes a transfer operation that is paused.}
\usage{
transferOperations.resume(ResumeTransferOperationRequest, name)
}
\arguments{
\item{ResumeTransferOperationRequest}{The \link{ResumeTransferOperationRequest} object to pass to this method}

\item{name}{The name of the transfer operation}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/storage/transfer}{Google Documentation}

Other ResumeTransferOperationRequest functions: \code{\link{ResumeTransferOperationRequest}}
}
