% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_objects.R
\name{TransferOptions}
\alias{TransferOptions}
\title{TransferOptions Object}
\usage{
TransferOptions(overwriteObjectsAlreadyExistingInSink = NULL,
  deleteObjectsUniqueInSink = NULL,
  deleteObjectsFromSourceAfterTransfer = NULL)
}
\arguments{
\item{overwriteObjectsAlreadyExistingInSink}{Whether overwriting objects that already exist in the sink is allowed}

\item{deleteObjectsUniqueInSink}{Whether objects that exist only in the sink should be deleted}

\item{deleteObjectsFromSourceAfterTransfer}{Whether objects should be deleted from the source after they are transferred to the sink}
}
\value{
TransferOptions object
}
\description{
TransferOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
TransferOptions uses three boolean parameters to define the actions to be performed on objects in a transfer.
}
