% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_objects.R
\name{TransferSpec}
\alias{TransferSpec}
\title{TransferSpec Object}
\usage{
TransferSpec(gcsDataSource = NULL, awsS3DataSource = NULL,
  httpDataSource = NULL, gcsDataSink = NULL, objectConditions = NULL,
  transferOptions = NULL)
}
\arguments{
\item{gcsDataSource}{A Google Cloud Storage data source}

\item{awsS3DataSource}{An AWS S3 data source}

\item{httpDataSource}{An HTTP URL data source}

\item{gcsDataSink}{A Google Cloud Storage data sink}

\item{objectConditions}{Only objects that satisfy these object conditions are included in the set of data source and data sink objects}

\item{transferOptions}{If the option `deleteObjectsUniqueInSink` is `true`, object conditions based on objects' `lastModificationTime` are ignored and do not exclude objects in a data source or a data sink}
}
\value{
TransferSpec object
}
\description{
TransferSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration for running a transfer.
}

