% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagetransfer_objects.R
\name{UpdateTransferJobRequest}
\alias{UpdateTransferJobRequest}
\title{UpdateTransferJobRequest Object}
\usage{
UpdateTransferJobRequest(projectId = NULL, transferJob = NULL,
  updateTransferJobFieldMask = NULL)
}
\arguments{
\item{projectId}{The ID of the Google Developers Console project that owns the job}

\item{transferJob}{The job to update}

\item{updateTransferJobFieldMask}{The field mask of the fields in `transferJob` that are to be updated in this request}
}
\value{
UpdateTransferJobRequest object
}
\description{
UpdateTransferJobRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request passed to UpdateTransferJob.
}
\seealso{
Other UpdateTransferJobRequest functions: \code{\link{transferJobs.patch}}
}

